1.add gcc clang target 2.PrimitiveExpression cast

+++ runtime/outcpp/System.Private.CoreLib/Internal/Win32/Registry.cpp
@@ -1,14 +1,15 @@
 #include "Registry-dep.h"
 
 #include <System.Private.CoreLib/Internal/Win32/RegistryKey-dep.h>
+#include <System.Private.CoreLib/System/Int32-dep.h>
 #include <System.Private.CoreLib/System/IntPtr-dep.h>
 
 namespace System::Private::CoreLib::Internal::Win32::RegistryNamespace {
 using namespace System;
 
 void Registry::cctor() {
-  CurrentUser = RegistryKey::in::OpenBaseKey((IntPtr)(-2147483647));
-  LocalMachine = RegistryKey::in::OpenBaseKey((IntPtr)(-2147483646));
+  CurrentUser = RegistryKey::in::OpenBaseKey((IntPtr)(Int32)(-2147483647));
+  LocalMachine = RegistryKey::in::OpenBaseKey((IntPtr)(Int32)(-2147483646));
 }
 
 } // namespace System::Private::CoreLib::Internal::Win32::RegistryNamespace
